feat: move into src-style structure, add database module with get prisma util, write first hello world code
