+++ package.json
@@ -5,9 +5,9 @@
     "author": "makerdao",
     "license": "AGPL-3",
     "scripts": {
-        "dev": "vite-node index.ts",
+        "dev": "vite-node src/index.ts",
         "postinstall": "prisma generate",
-       "lint": "eslint \"./**/*.{ts,tsx}\" --max-warnings=0",
+        "lint": "eslint \"./**/*.{ts,tsx}\" --max-warnings=0",
         "typecheck": "tsc --noEmit"
     },
     "dependencies": {
@@ -16,6 +16,8 @@
     },
     "devDependencies": {
         "@types/node": "^18.14.2",
+        "@typescript-eslint/eslint-plugin": "^5.54.0",
+        "@typescript-eslint/parser": "^5.54.0",
         "eslint": "^8.35.0",
         "eslint-config-airbnb-base": "^15.0.0",
         "prisma": "^4.11.0",
