+++ src/database/index.ts
@@ -1,0 +1,11 @@
+import { PrismaClient } from '@prisma/client';
+
+let prisma: PrismaClient;
+
+export const getPrisma = () => {
+  if (!prisma) {
+    prisma = new PrismaClient();
+  }
+
+  return prisma;
+};
